facemap_flask_backend/requirements.txt
backcall==0.2.0
bidict==0.21.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
conda==4.3.16
decorator==5.0.7
Flask==1.1.2
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==2.10
install==1.3.4
ipython==7.23.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
MarkupSafe==1.1.1
matplotlib-inline==0.1.2
numpy==1.20.2
opencv-contrib-python==4.1.2.30
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycosat==0.6.3
Pygments==2.9.0
python-engineio==4.1.0
python-socketio==5.2.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.15.0
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
websocket-client==0.58.0
Werkzeug==1.0.1